Comment out the call to tfw_http_add_or_append_forwarded_for()
tfw_http_add_or_append_forwarded_for() is currently broken, which
leads to system crash (issue #47). This functionality is not crucial
at this time, so just comment it out until it is fixed, to allow for
progress in other parts of the code.

Relax the compilation options in tempesta_fw/Makefile to allow the
compilation pass okay with warnings. That became a necessity when
a call to tfw_http_add_or_append_forwarded_for() was commented out.