rcl: add limits for sizes of uri/headers/body of a HTTP request
The change enhances the req_conn_limit module with hooks that limit
the maximum length for uri/headers/body in a HTTP request.

The new sysctl settings are available:
 - http_uri_len - the limit for the URI length
 - http_field_len - the limit for the length of a single header field
 - http_body_len - the limit for the body of a HTTP request

All these limits are checked when for every received chunk of data.
That allows to drop a HTTP request just after it exceeds a limit without
waiting until it is fully received.