Detach an SKB from the receive socket in a correct way (Issue #60)
When an SKB is put on the socket's receive queue it's accounted for.
So when we remove it from the queue we need to adjust for that in
the accounting department. In a regular course of things that happens
when an SKB is freed. An SKB destruct() function that assigned to it
by the socket takes care of that. Orphaning an SKB from the socket
triggers the same action without actual freeing of an SKB.