Initial attempt at making Synchronous Sockets (SS) simpler
Make SS simpler by removing lots of Linux socket code from SS.
Leave that job to Linux, it does it well. When a socket leaves
the ESTABLISHED state, we release all Tempesta resources linked
to a socket, and do not take part in any work or actions that
are needed to have the socket closed. Linux does the usual job
of closing the TCP connection and related sockets in a correct way.

This DOES NOT solve the problem of a deadlock (issue #61).