http_parser: accumulate TfwStr chunks with  __FIELD_OPEN()/__FIELD_CLOSE()
Related issue: #72 - bad handling of fragmented HTTP messages

The main problem with this issue is the string data (the TfwStr object).
In certain places, you have save strings (host, uri, headers, body) into a
TfwMsg object. In order to do this, you need to remember a start position
in a FSM state, and refer back to it later when an end position is known
in another FSM state. But the problem is you can have a "context switch"
between these two states and then your start position is not valid anymore.

This commit adds the macros that helpes you to save the issue.
You just mark the start position with the __FIELD_OPEN() and end positon
with the __FIELD_CLOSE() and the macros save all the string data automaically
with splitting it into chunks when the FSM interrupts its execution.

ALso, this commit integrates this macro with existing code that parses:
URI/Host and header fields.