http_parser: fix __FSM_MOVE(0) in __HDR_PARSE_VAL() helper
The __HDR_PARSE_VAL() uses a helper function that parses the value.
The function returns length of the parsed value, and the macro does
__FSM_MOVE_n(current_state, parsed_len).

The problem is the header value length may be less than the parsed data
because there is CRLF behind the value, and it is parsed by the helper
function tool.

A particular example is the header: "Host: localhost\r\n".
When it is split into two chunks: "Host: localhost" and "\r\n", and the
parser reaches the second chunk, the __req_parse_host() helper returns
value_length=0 because there is no value in the "\r\n".
Basically that leads to an infinite loop because of __FSM_MOVE_n(0).

This commit fixes the CSTR_POSTPONE by doing the __FSM_MOVE_n() to a
correct position.

Related issue: #72.