http_parser: fix CSTR_POSTPONE'd chunk length in __HDR_PARSE_VAL()
The __HDR_PARSE_VAL() uses a helper parsing function with an inferior FSM.
The function parses not only the header value, but also EOL after it.

The problem arises when the data buffer ends at this EOL.
The function returns CSTR_POSTPONE, but doesn't signal in any way that the
value is fully parsed, and the __HDR_PARSE_VAL() treats the EOL as a part
of the value.

The commit adds the check that detects the parsed value and closes the header
at that position. However, the solution is not elegant.

Related issue: #72